+++ frontend/src/utils.ts
@@ -29,3 +29,22 @@ export const getWeatherData = async (stationId?: number) => {
   const res = await fetch(url.toString());
   return res.json();
 };
+
+export const getHealthImpact = (minutes: number) => {
+  // Cycle commuting was associated with a lower risk of CVD, cancer, and all cause mortality.
+  // Link to the study: [https://www.bmj.com/content/357/bmj.j1456?tab=related#datasupp]
+
+  // regular cycling cut the risk of death from any cause by 41%, the incidence of cancer by 45% and heart disease by 46%.
+  // Link [https://www.bbc.com/news/health-39641122]
+
+  // Each week, adults should move briskly for at least 150 minutes
+
+  return {
+    percentOfRecommendedActivity: Math.round((minutes / 150 / 7) * 100),
+    generalHealthBenefits: [
+      "cut the risk of death from any cause by 41%",
+      "decrease the incidence of cancer by 45%",
+      "decrease the incidence of heart disease by 46%",
+    ],
+  };
+};
